Feature for manage categories in lista view
Added the changes on models for include new column. Added in views the functions por all the actions that are needed for the table categories allowing the administers to do changes if needed. Added a new form based on categories. Added the path on urls to view the pages and added in lista.html all the js script and modal for categories management.